+++ src/js/page-background.js
@@ -1,0 +1,1 @@
+TrelloChecker.run(true);